Add new state prop to show votes in input box
When fixing the manual input bug, another emerged which caused the
value of the input box to not update whenever the intended input would
have led to an excess vote error.

This commit fixes that bug by creating a new state prop which updates
the value shown in the input box, with the former state being used to
update it left to perform updates needed for conditional rendering only.